Logistics: move into src folder with integration
