--- Logistics/index.js
@@ -1,6 +1,0 @@
-// This file will hopefully be the home base, I suppose,
-// where the game is initiallized, and brought together as one.
-
-import TestShip from "../Sandbox/Scripts/Ship/ColonyShipTester.js";
-let testShip = new TestShip();
-testShip.start();
