+++ index.html
@@ -5,22 +5,21 @@
 	<meta charset="UTF-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="Logistics/style.css">
+    <link rel="stylesheet" href="style.css">
 	<title>Shape Ship Collaboration Activity</title>
 </head>
 <body>
     <!-- Canvas initialization -->
     <section class="canvasHolder">
-        <!-- These are created with JavaScript (screen.js):
-            <canvas id="background"></canvas>
-            <canvas id="missiles"></canvas>
-            <canvas id="objects"></canvas>
-            <canvas id="thrusters"></canvas>
-            <canvas id="ships"></canvas>
-            <canvas id="items"></canvas>
+        <!-- These are created with JavaScript (src/index.js):
+        <canvas id="background"></canvas>
+        <canvas id="missiles"></canvas>
+        <canvas id="objects"></canvas>
+        <canvas id="thrusters"></canvas>
+        <canvas id="ships"></canvas>
+        <canvas id="items"></canvas>
         -->
     </section>
+    <script src="src/index.js" type="module"></script>
 </body>
-<script src="Logistics/screen.js" type="module"></script>
-<script src="Logistics/index.js" type="module"></script>
 </html>
