+++ style.css
@@ -1,0 +1,24 @@
+@font-face {
+    font-family: NotoSans-Black;
+    src: url("../Sandbox/Fonts/NotoSans-Black.ttf");
+}
+
+body {
+    --unit: 30px; /* screen.js updates this. */
+    --border: 6px; /* screen.js updates this. */
+    --width: 32; /* screen.js updates this. */
+    --height: 24; /* screen.js updates this. */
+    font-family: NotoSans-Black;
+    overflow: hidden;
+    background-color: black;
+}
+
+canvas {
+    position: absolute;
+    /* Math to center the canvases in the viewing window. */
+    left: calc(50% - var(--width) / 2);
+    top: calc(50% - var(--height) / 2);
+    width: calc(var(--width) * var(--unit));
+    height: calc(var(--height) * var(--unit));
+    outline: var(--border) solid rgba(50, 50, 50);
+}
